MCSC-1127 - Spike for Implementing Rounding Logic

+++ app/models/CalculationResults/OutOfDatesTaxYearsCalculation.scala
@@ -21,15 +21,15 @@ import play.api.libs.json.{Reads, Writes, __}
 
 case class OutOfDatesTaxYearsCalculation(
   period: Period,
-  directCompensation: Int,
-  indirectCompensation: Int,
-  chargePaidByMember: Int,
-  chargePaidBySchemes: Int,
-  revisedChargableAmountBeforeTaxRate: Int,
-  revisedChargableAmountAfterTaxRate: Int,
-  unusedAnnualAllowance: Int,
+  directCompensation: Double,
+  indirectCompensation: Double,
+  chargePaidByMember: Double,
+  chargePaidBySchemes: Double,
+  revisedChargableAmountBeforeTaxRate: Double,
+  revisedChargableAmountAfterTaxRate: Double,
+  unusedAnnualAllowance: Double,
   taxYearSchemes: List[OutOfDatesTaxYearSchemeCalculation],
-  adjustedCompensation: Option[Int]
+  adjustedCompensation: Option[Double]
 )
 
 object OutOfDatesTaxYearsCalculation {
@@ -40,15 +40,15 @@ object OutOfDatesTaxYearsCalculation {
 
     (
       (__ \ "period").read[Period] and
-        (__ \ "directCompensation").read[Int] and
-        (__ \ "indirectCompensation").read[Int] and
-        (__ \ "chargePaidByMember").read[Int] and
-        (__ \ "chargePaidBySchemes").read[Int] and
-        (__ \ "revisedChargableAmountBeforeTaxRate").read[Int] and
-        (__ \ "revisedChargableAmountAfterTaxRate").read[Int] and
-        (__ \ "unusedAnnualAllowance").read[Int] and
+        (__ \ "directCompensation").read[Double] and
+        (__ \ "indirectCompensation").read[Double] and
+        (__ \ "chargePaidByMember").read[Double] and
+        (__ \ "chargePaidBySchemes").read[Double] and
+        (__ \ "revisedChargableAmountBeforeTaxRate").read[Double] and
+        (__ \ "revisedChargableAmountAfterTaxRate").read[Double] and
+        (__ \ "unusedAnnualAllowance").read[Double] and
         (__ \ "taxYearSchemes").read[List[OutOfDatesTaxYearSchemeCalculation]] and
-        (__ \ "adjustedCompensation").readNullable[Int]
+        (__ \ "adjustedCompensation").readNullable[Double]
     )(OutOfDatesTaxYearsCalculation.apply _)
 
   }
@@ -59,15 +59,15 @@ object OutOfDatesTaxYearsCalculation {
 
     (
       (__ \ "period").write[Period] and
-        (__ \ "directCompensation").write[Int] and
-        (__ \ "indirectCompensation").write[Int] and
-        (__ \ "chargePaidByMember").write[Int] and
-        (__ \ "chargePaidBySchemes").write[Int] and
-        (__ \ "revisedChargableAmountBeforeTaxRate").write[Int] and
-        (__ \ "revisedChargableAmountAfterTaxRate").write[Int] and
-        (__ \ "unusedAnnualAllowance").write[Int] and
+        (__ \ "directCompensation").write[Double] and
+        (__ \ "indirectCompensation").write[Double] and
+        (__ \ "chargePaidByMember").write[Double] and
+        (__ \ "chargePaidBySchemes").write[Double] and
+        (__ \ "revisedChargableAmountBeforeTaxRate").write[Double] and
+        (__ \ "revisedChargableAmountAfterTaxRate").write[Double] and
+        (__ \ "unusedAnnualAllowance").write[Double] and
         (__ \ "taxYearSchemes").write[List[OutOfDatesTaxYearSchemeCalculation]] and
-        (__ \ "adjustedCompensation").writeNullable[Int]
+        (__ \ "adjustedCompensation").writeNullable[Double]
     )(a =>
       (
         a.period,
